WIP: Related to #197. Enabling different views for functions.
In order to implement different behaviour when displaying results for
special functions (needed for tip radius in power spectrum and contact
mechanics), "card_view_flavors" have been introduced.

At the moment, there are two flavors: "simple" and "plot".

"simple": Just the dict of the result is displayed in the view
"plot": A 2D plot of all series in the results, this is now
        set for all analysis functions.

Later also : "psd plot", "contact mechanics"

The flavor is set with the decorator "@analysis_function()".

Additionally for each flavor there are "template flavors":
Currently "list" and "detail" are used. They are given as GET arguments
in the AJAX calls for the cards content. Depending on the value,
the approriate template is selected. So we can have multiple templates
for the same view. This is needed to display the card in another
way when "openend" in the search results, first "list", then "detail".