ENH: Added organizations for saving available plugins.
Each organization is related to one specific group
(could be more in future) and each group can have zero to
many users as members.
Each organization has a comma-separated list of names
of available plugins. Each user who ist member of an organization
via a group has access to all plugins which are available
through these organizations.

Currently the function names presented on the Analysis page
are filtered by the plugins avaiable.

What's missing is that the views of a plugin are only accessible
if the user belongs to an organization which has access to
a plugin. This can be done by automatic wrapping of the plugin
views with a permission decorator (already present as stub
in a comment).